backend/requirements.txt
# Web Framework
alembic
fastapi
bcrypt==3.2.2
passlib
psycopg2-binary
python-jose[cryptography]
python-multipart
sqlalchemy
uvicorn[standard]
pydantic[email]